+++ lib/help_desk/help_desk_main/view/desktop/replyChannel/description_desktop.dart
