+++ lib/help_desk/help_desk_main/view/mobile/mobile_widget.dart
@@ -6,40 +6,25 @@ import 'package:senior_project/assets/font_style.dart';
 import 'package:senior_project/core/datasource/firebase_services.dart';
 import 'package:senior_project/core/view_model/app_view_model.dart';
 import 'package:senior_project/help_desk/help_desk_main/view/widget/loader_status.dart';
-import 'package:senior_project/help_desk/help_desk_main/view/widget/mobile/create_task.dart';
-import 'package:senior_project/help_desk/help_desk_main/view/widget/mobile/setting_pop_up.dart';
-import 'package:senior_project/help_desk/help_desk_main/view/widget/mobile/task_card.dart';
+import 'package:senior_project/help_desk/help_desk_main/view/mobile/create_task.dart';
+import 'package:senior_project/help_desk/help_desk_main/view/mobile/setting_pop_up.dart';
+import 'package:senior_project/help_desk/help_desk_main/view/mobile/ticket_card.dart';
 import 'package:senior_project/help_desk/help_desk_main/view_model/help_desk_view_model.dart';
 
-Stream? query(String id, int type) {
+Stream? query(String id, int type, bool isAdmin) {
   final FirebaseServices service = FirebaseServices("ticket");
- if (id.isEmpty) {
-    switch (type) {
-      case 0:
-        return service.listenToDocument();
-      case 1:
-        return service.listenToDocumentByKeyValuePair(["status"], [0]);
-      case 2: 
-        return service.listenToDocumentByKeyValuePair(["status"], [1]);
-      case 3:
-        return service.listenToDocumentByKeyValuePair(["status"], [2]);
-      default:
-        return null;
-    }
-  } else {
-    switch (type) {
-      case 0:
-        return service.listenToDocumentByKeyValuePair(["ownerId"], [id]);
-      case 1:
-        return service.listenToDocumentByKeyValuePair(["ownerId", "status"], [id, 0]);
-      case 2:
-        return service.listenToDocumentByKeyValuePair(["ownerId", "status"], [id, 1]);
-      case 3:
-        return service.listenToDocumentByKeyValuePair(["ownerId", "status"], [id, 2]);
-      default:
-        return null;
-    }
+  if (type == 0) {
+    return service.listenToDocumentByKeyValuePair(
+    [isAdmin ? "adminId" : "ownerId"],
+    [id], 
+    descending: true
+  );
   }
+  return service.listenToDocumentByKeyValuePair(
+    [isAdmin ? "adminId" : "ownerId", "status"],
+    [id, type-1], 
+    descending: true
+  );
 }
 
 class MobileWidget extends StatefulWidget {
@@ -85,7 +70,7 @@ class _MobileWidgetState extends State<MobileWidget> {
   List<Widget> generateContent(List<Map<String, dynamic>> content) {
     List<Widget> list = [];
     for (int i = 0; i < content.length; i++) {
-      list.add(TaskCard(detail: content[i]));
+      list.add(TicketCard(detail: content[i]));
     }
     return list;
   }
@@ -95,7 +80,7 @@ class _MobileWidgetState extends State<MobileWidget> {
     int tagBarSelected = context.watch<HelpDeskViewModel>().getSelectedMobileMenu();
     String id = context.watch<AppViewModel>().app.getUser.getId;
     context.read<HelpDeskViewModel>().cleanModel();
-    _stream = query(widget.isAdmin ? "" : id, tagBarSelected);
+    _stream = query(id, tagBarSelected, widget.isAdmin);
     super.didChangeDependencies();
   }
