+++ lib/help_desk/help_desk_main/view/page/help_desk_main_view.dart
@@ -5,9 +5,12 @@ import 'package:senior_project/assets/font_style.dart';
 import 'package:senior_project/core/template_desktop/view/page/template_desktop.dart';
 import 'package:senior_project/core/template_mobile/view/template_menu_mobile.dart';
 import 'package:senior_project/core/view_model/app_view_model.dart';
-import 'package:senior_project/help_desk/help_desk_main/view/widget/desktop/body.dart';
-import 'package:senior_project/help_desk/help_desk_main/view/widget/desktop/header.dart';
-import 'package:senior_project/help_desk/help_desk_main/view/widget/mobile/mobile_widget.dart';
+import 'package:senior_project/help_desk/help_desk_main/view/desktop/body.dart';
+import 'package:senior_project/help_desk/help_desk_main/view/desktop/header.dart';
+import 'package:senior_project/help_desk/help_desk_main/view/mobile/mobile_widget.dart';
+import 'package:senior_project/help_desk/help_desk_main/view_model/help_desk_view_model.dart';
+import 'package:senior_project/help_desk/help_desk_reply/view/page/help_desk_reply_page.dart';
+import 'package:senior_project/help_desk/help_desk_reply/view_model/reply_channel_view_model.dart';
 
 class HelpDeskMainView extends StatelessWidget {
   final bool isAdmin;
@@ -19,8 +22,25 @@ class HelpDeskMainView extends StatelessWidget {
     double screenHeight = MediaQuery.of(context).size.height;
     int? role = context.watch<AppViewModel>().app.getUser.getRole;
     bool isLogin = context.watch<AppViewModel>().isLogin;
+    // bool isShowMessagePage = context.watch<HelpDeskViewModel>().getIsShowMessagePage;
+    // Map<String, dynamic> messageDetail = context.watch<ReplyChannelViewModel>().getTaskData;
 
     if (isMobileSite) {
+      // if (isShowMessagePage) {
+      //   Navigator.push(context, MaterialPageRoute(builder: (context) {
+      //     return HelpDeskReplyPage(
+      //     docId: messageDetail["docId"],
+      //     id: messageDetail["id"],
+      //     title: messageDetail["title"],
+      //     detail: messageDetail["detail"],
+      //     priority: messageDetail["priority"],
+      //     status: messageDetail["status"],
+      //     category: messageDetail["category"],
+      //     time: messageDetail["time"]
+      //   );
+      //   }));
+        
+      // }
       return TemplateMenuMobile(
           content: Builder(
             builder: (context) {
@@ -88,10 +108,5 @@ class HelpDeskMainView extends StatelessWidget {
             );
           }
         ));
-    // if (isAdmin) {
-    //   return const HelpDeskAdminPage();
-    // } else {
-    //   return const HelpDeskMainDesktopWidget();
-    // }
   }
 }
