+++ lib/help_desk/help_desk_reply/view/page/help_desk_reply_page.dart
@@ -1,25 +1,28 @@
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 import 'package:senior_project/core/view_model/app_view_model.dart';
-import 'package:senior_project/help_desk/help_desk_reply/view/widget/desktop/help_desk_reply_desktop.dart';
 import 'package:senior_project/help_desk/help_desk_reply/view/widget/mobile/help_desk_reply_mobile.dart';
 import 'package:senior_project/help_desk/help_desk_reply/view_model/reply_channel_view_model.dart';
 
 class HelpDeskReplyPage extends StatefulWidget {
   final String docId;
-  final String taskId;
-  final String taskTitle;
-  final String taskDetail;
+  final String id;
+  final String title;
+  final String detail;
   final int priority;
   final int status;
+  final String category;
+  final DateTime time;
   const HelpDeskReplyPage({
     super.key, 
     required this.docId,
-    required this.taskId,
-    required this.taskTitle,
-    required this.taskDetail,
+    required this.id,
+    required this.title,
+    required this.detail,
     required this.priority,
-    required this.status
+    required this.status,
+    required this.category,
+    required this.time
   });
 
   @override
@@ -32,16 +35,18 @@ class _HelpDeskReplyPageState extends State<HelpDeskReplyPage> {
     bool isMobile = context.watch<AppViewModel>().getMobileSiteState(MediaQuery.of(context).size.width);
     context.read<ReplyChannelViewModel>().setTaskData = {
       "docId": widget.docId,
-      "id": widget.taskId,
-      "title": widget.taskTitle,
-      "detail": widget.taskDetail,
+      "id": widget.id,
+      "title": widget.title,
+      "detail": widget.detail,
       "priority": widget.priority,
-      "status": widget.status
+      "status": widget.status,
+      "category": widget.category,
+      "time": widget.time
     };
 
     if (isMobile) {
       return const HelpDeskReplyMobile();
     }
-    return const HelpDeskReplyDesktop();
+    return Container();
   }
 }
